+++ app/tps_counter_test.go
@@ -70,5 +70,6 @@ func (wl *writerLogger) Info(msg string, keyVals ...interface{}) {
 	defer wl.mu.Unlock()
 
 	wl.nTotalTPS += keyVals[1].(float64)
+	//nolint
 	fmt.Fprintf(wl.w, msg+" "+fmt.Sprintf("%s %.2f\n", keyVals[0], keyVals[1]))
 }