+++ precompiles/auctions/IAuctions.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.0;
 import "../common/Types.sol";
 
 /// @dev The IAuctions contract's address.
-address constant AUCTIONS_PRECOMPILE_ADDRESS = 0x0000000000000000000000000000000000000805;
+address constant AUCTIONS_PRECOMPILE_ADDRESS = 0x0000000000000000000000000000000000000900;
 
 /// @dev The ITokenFactory contract's instance.
 IAuctions constant AUCTIONS_CONTRACT = IAuctions(AUCTIONS_PRECOMPILE_ADDRESS);
@@ -20,24 +20,27 @@ struct AuctionInfo {
 /// @author Evmos Team
 /// @title Auctions Precompiled Contract
 /// @dev The interface through which solidity contracts interacts with the burn auctions module
-/// @custom:address 0x0000000000000000000000000000000000000805
+/// @custom:address 0x0000000000000000000000000000000000000900
 interface IAuctions {
     // @dev Event emitted when a new bid is made for the basket of Coins.
     // @param sender - the hex address of the sender of the bid.
+    // @param round - the round of the auction.
     // @param amount - the amount in Evmos tokens for the bid.
-    event Bid(address indexed sender, uint256 amount);
+    event Bid(address indexed sender, uint64 indexed round, uint256 amount);
 
     // @dev Event emitted when a new Coin deposit is made for the following burn auction.
     // @param sender - the hex address of the sender of the Coins.
+    // @param round - the round of the auction.
     // @param denom - the denom of the Coin being sent.
     // @param amount - the amount of the Coin being sent.
-    event CoinDeposit(address indexed sender, string denom, uint256 amount);
+    event CoinDeposit(address indexed sender, uint64 indexed round, address denom, uint256 amount);
 
     // @dev Event emitted when a burn auction ends.
     // @param winner - the hex address of the auction winner.
+    // param round - the round of the auction.
     // @param coins - the Coins won in the auction.
     // @param burned - the amount of tokens burned on the auction.
-    event AuctionEnd(address indexed winner, Coin[] coins, uint256 burned);
+    event AuctionEnd(address indexed winner, uint64 indexed round, Coin[] coins, uint256 burned);
 
     // @dev Creates a bid for the basket of assets in the current auction.
     // @dev param sender - the hex address of the sender.