+++ precompiles/auctions/auctions.go
@@ -21,7 +21,7 @@ import (
 
 var _ vm.PrecompiledContract = &Precompile{}
 
-const PrecompileAddress string = "0x0000000000000000000000000000000000000805"
+const PrecompileAddress string = "0x0000000000000000000000000000000000000900"
 
 // Embed abi json file to the executable binary. Needed when importing as dependency.
 //