+++ precompiles/authorization/types.go
@@ -74,6 +74,7 @@ func CheckApprovalArgs(args []interface{}, denom string) (common.Address, *sdk.C
 		return common.Address{}, nil, nil, fmt.Errorf(ErrInvalidMethods, args[2])
 	}
 	if len(typeURLs) == 0 {
+		//nolint
 		return common.Address{}, nil, nil, fmt.Errorf(ErrEmptyMethods)
 	}
 	if slices.Contains(typeURLs, "") {
@@ -191,6 +192,7 @@ func validateMsgTypes(arg interface{}) ([]string, error) {
 		return nil, fmt.Errorf(ErrInvalidMethods, arg)
 	}
 	if len(typeURLs) == 0 {
+		//nolint
 		return nil, fmt.Errorf(ErrEmptyMethods)
 	}