+++ precompiles/ics20/types.go
@@ -108,11 +108,13 @@ func NewMsgTransfer(method *abi.Method, args []interface{}) (*transfertypes.MsgT
 
 	sourcePort, ok := args[0].(string)
 	if !ok {
+		//nolint
 		return nil, common.Address{}, fmt.Errorf(ErrInvalidSourcePort)
 	}
 
 	sourceChannel, ok := args[1].(string)
 	if !ok {
+		//nolint
 		return nil, common.Address{}, fmt.Errorf(ErrInvalidSourceChannel)
 	}
 
@@ -273,11 +275,13 @@ func checkAllowanceArgs(args []interface{}) (common.Address, string, string, str
 
 	sourcePort, ok := args[1].(string)
 	if !ok {
+		//nolint
 		return common.Address{}, "", "", "", nil, fmt.Errorf(ErrInvalidSourcePort)
 	}
 
 	sourceChannel, ok := args[2].(string)
 	if !ok {
+		//nolint
 		return common.Address{}, "", "", "", nil, fmt.Errorf(ErrInvalidSourceChannel)
 	}