+++ x/auctions/keeper/events.go
@@ -21,39 +21,7 @@ import (
 	evmtypes "github.com/evmos/evmos/v19/x/evm/types"
 )
 
-//// EmitRoundFinished emits an event for the RoundFinished event.
-//func EmitRoundFinished(ctx sdk.Context, evmKeeper evmkeeper.Keeper, stateDB vm.StateDB, roundID *big.Int) error {
-//	// Get the precompile instance
-//	evmParams := evmKeeper.GetParams(ctx)
-//	precompile, found, err := evmKeeper.GetStaticPrecompileInstance(&evmParams, common.HexToAddress(auctionsprecompile.PrecompileAddress))
-//	if err != nil || !found {
-//		return err
-//	}
-//
-//	// Prepare the events
-//	p := precompile.(auctionsprecompile.Precompile)
-//	event := p.Events[auctionsprecompile.EventTypeRoundFinished]
-//	topics := make([]common.Hash, 1)
-//
-//	// The first topic is always the signature of the event.
-//	topics[0] = event.ID
-//
-//	// Pack the arguments to be used as the Data field
-//	arguments := abi.Arguments{event.Inputs[0]}
-//	packed, err := arguments.Pack(roundID)
-//	if err != nil {
-//		return err
-//	}
-//
-//	stateDB.AddLog(&ethtypes.Log{
-//		Address:     p.Address(),
-//		Topics:      topics,
-//		Data:        packed,
-//		BlockNumber: uint64(ctx.BlockHeight()),
-//	})
-//
-//	return nil
-//}
+const PrecompileAddress = "0x0000000000000000000000000000000000000900"
 
 var EndAuctionEventABI = abi.Event{
 	Name:      "AuctionEnd",
@@ -62,37 +30,42 @@ var EndAuctionEventABI = abi.Event{
 	Inputs: abi.Arguments{
 		abi.Argument{
 			Name:    "winner",
-			Type:    abi.Type{Size: 20, T: 7},
+			Type:    abi.Type{Size: 20, T: abi.AddressTy},
+			Indexed: true,
+		},
+		abi.Argument{
+			Name:    "round",
+			Type:    abi.Type{Size: 64, T: abi.UintTy},
 			Indexed: true,
 		},
 		abi.Argument{
 			Name: "coins",
 			Type: abi.Type{
 				Elem: &abi.Type{
-					T:            6,
+					T:            abi.TupleTy,
 					TupleRawName: "Coin",
 					TupleElems: []*abi.Type{
-						{T: 3},
-						{Size: 256, T: 1},
+						{T: abi.StringTy},
+						{Size: 256, T: abi.UintTy},
 					},
 					TupleRawNames: []string{"denom", "amount"},
 					TupleType:     reflect.TypeOf(contracttypes.Coin{}),
 				},
-				T: 4,
+				T: abi.SliceTy,
 			},
 		},
 		abi.Argument{
 			Name: "burned",
-			Type: abi.Type{Size: 256, T: 1},
+			Type: abi.Type{Size: 256, T: abi.UintTy},
 		},
 	},
-	Sig: "AuctionEnd(address,(string,uint256)[],uint256)",
-	ID:  crypto.Keccak256Hash([]byte("AuctionEnd(address,(string,uint256)[],uint256)")),
+	Sig: "AuctionEnd(address,uint64,(string,uint256)[],uint256)",
+	ID:  crypto.Keccak256Hash([]byte("AuctionEnd(address,uint64,(string,uint256)[],uint256)")),
 }
 
 // EmitAuctionEndEvent emits an event as an ethereum tx log to be able to filter
 // it via the JSON-RPC
-func EmitAuctionEndEvent(ctx sdk.Context, winner sdk.AccAddress, coins sdk.Coins, burnedAmt math.Int) error {
+func EmitAuctionEndEvent(ctx sdk.Context, winner sdk.AccAddress, round uint64, coins sdk.Coins, burnedAmt math.Int) error {
 	bidWinnerHexAddr := common.BytesToAddress(winner.Bytes())
 
 	// event topics
@@ -101,13 +74,20 @@ func EmitAuctionEndEvent(ctx sdk.Context, winner sdk.AccAddress, coins sdk.Coins
 		return errorsmod.Wrapf(err, "failed make log topic")
 	}
 
-	// index the bidWinner address
+	roundTopic, err := contractutils.MakeTopic(round)
+	if err != nil {
+		return errorsmod.Wrapf(err, "failed make log topic")
+	}
+
+	// index the bidWinner address and round
 	topics := []common.Hash{
 		EndAuctionEventABI.ID,
 		winnerTopic,
+		roundTopic,
 	}
+
 	// Pack the arguments to be used as the Data field
-	arguments := abi.Arguments{EndAuctionEventABI.Inputs[1], EndAuctionEventABI.Inputs[2]}
+	arguments := abi.Arguments{EndAuctionEventABI.Inputs[2], EndAuctionEventABI.Inputs[3]}
 
 	// parse coins to use big int instead of sdkmath.Int
 	eventCoins := contracttypes.NewCoinsResponse(coins)
@@ -118,7 +98,7 @@ func EmitAuctionEndEvent(ctx sdk.Context, winner sdk.AccAddress, coins sdk.Coins
 	}
 
 	ethLog := &ethtypes.Log{
-		Address:     common.HexToAddress("0x0000000000000000000000000000000000000805"), // ?? set the auctions precompile address in the log. Or should we use the auctions mod address?  TODO: get it from a constaant instead of hardcoded
+		Address:     common.HexToAddress(PrecompileAddress),
 		Topics:      topics,
 		Data:        packed,
 		BlockNumber: uint64(ctx.BlockHeight()),